+++ packages/extension-link/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.35](https://github.com/ueberdosis/tiptap/compare/@tiptap/extension-link@2.0.0-beta.34...@tiptap/extension-link@2.0.0-beta.35) (2022-01-07)
+
+
+### Bug Fixes
+
+* allow to remove autolinks with unsetLink ([ea10ffb](https://github.com/ueberdosis/tiptap/commit/ea10ffbc6a0f447c33680dcc6f6dc2fa6337c4c2))
+
+
+
+
+
 # [2.0.0-beta.34](https://github.com/ueberdosis/tiptap/compare/@tiptap/extension-link@2.0.0-beta.33...@tiptap/extension-link@2.0.0-beta.34) (2022-01-04)
 
 **Note:** Version bump only for package @tiptap/extension-link