+++ packages/starter-kit/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.171](https://github.com/ueberdosis/tiptap/compare/@tiptap/starter-kit@2.0.0-beta.170...@tiptap/starter-kit@2.0.0-beta.171) (2022-01-13)
+
+**Note:** Version bump only for package @tiptap/starter-kit
+
+
+
+
+
+# [2.0.0-beta.170](https://github.com/ueberdosis/tiptap/compare/@tiptap/starter-kit@2.0.0-beta.169...@tiptap/starter-kit@2.0.0-beta.170) (2022-01-11)
+
+**Note:** Version bump only for package @tiptap/starter-kit
+
+
+
+
+
+# [2.0.0-beta.169](https://github.com/ueberdosis/tiptap/compare/@tiptap/starter-kit@2.0.0-beta.168...@tiptap/starter-kit@2.0.0-beta.169) (2022-01-10)
+
+**Note:** Version bump only for package @tiptap/starter-kit
+
+
+
+
+
+# [2.0.0-beta.168](https://github.com/ueberdosis/tiptap/compare/@tiptap/starter-kit@2.0.0-beta.167...@tiptap/starter-kit@2.0.0-beta.168) (2022-01-05)
+
+**Note:** Version bump only for package @tiptap/starter-kit
+
+
+
+
+
+# [2.0.0-beta.167](https://github.com/ueberdosis/tiptap/compare/@tiptap/starter-kit@2.0.0-beta.166...@tiptap/starter-kit@2.0.0-beta.167) (2022-01-05)
+
+**Note:** Version bump only for package @tiptap/starter-kit
+
+
+
+
+
+# [2.0.0-beta.166](https://github.com/ueberdosis/tiptap/compare/@tiptap/starter-kit@2.0.0-beta.165...@tiptap/starter-kit@2.0.0-beta.166) (2022-01-05)
+
+**Note:** Version bump only for package @tiptap/starter-kit
+
+
+
+
+
 # [2.0.0-beta.165](https://github.com/ueberdosis/tiptap/compare/@tiptap/starter-kit@2.0.0-beta.164...@tiptap/starter-kit@2.0.0-beta.165) (2022-01-04)
 
 **Note:** Version bump only for package @tiptap/starter-kit